WIP: Add graphical display; move out of main.c
