+++ src/range_live.h
@@ -1,0 +1,3 @@
+#pragma once
+
+void display_live_ranges();
